requirements.txt
streamlit==1.28.1
torch==2.0.0
torchvision==0.15.0
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
matplotlib==3.7.2
scipy==1.11.1
scikit-image==0.21.0
